Continue monitoring renaming
